Break composer.lock
Heroku-WP is a template for running WP on Heroku. Forking this repo or
branching from the template branch results in lots of messy changes and merge
conflicts with the composer.lock file which should be rebuilt upon merge
anyway. To make merges and histories cleaner just commit a broken composer.lock
file and force everyone to generate their own from the start.